requirements.txt
requests>=2.31.0
geopandas>=0.14.0
shapely>=2.0.0
rapidfuzz>=3.6.0
python-slugify>=8.0.0
pandas>=2.2.0
SPARQLWrapper>=2.0.0
tqdm>=4.66.0
